move whenEnabled to config (same role as isEnabled)
