+++ functions/src/process-email-queue.ts
@@ -2,9 +2,8 @@ import * as functions from "firebase-functions";
 import { firestore } from "firebase-admin";
 import { DocumentSnapshot } from "firebase-functions/lib/providers/firestore";
 import * as mailgun from "mailgun-js";
-import { Config, isEnabled } from "./config";
+import { Config, onlyWhenEnabled } from "./config";
 import { ensureCalledOnce } from "./ensure-called-once";
-import { whenEnabled } from "./when-enabled";
 
 const config = functions.config() as Config;
 const mailgunClient = mailgun({
@@ -60,7 +59,7 @@ const processEmail = async (emailSnapshot: DocumentSnapshot) => {
 export const processEmailQueue = functions.firestore
   .document(`${EMAILS_TO_SEND_COLLECTION_NAME}/{id}`)
   .onCreate(
-    whenEnabled(
+    onlyWhenEnabled(
       config.features.emails,
       ensureCalledOnce(db, "process-email-queue", processEmail)
     )
